+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add support for nOS-V progressing events VP{pra}.
 - Add breakdown model for nOS-V.
+- New API to manage stream metadata `ovni_attr_*()`.
 
 ## [1.9.1] - 2024-05-10